ocr_service/requirements.txt
paddleocr>=2.7.0,<3.0.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
requests==2.31.0
